+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2024-03-05
+Features
+- Add slack prefix input
+
 # 2021-04-01
 
 Features